Fix CI issues: formatting, clippy warnings, and unused imports
Resolves all GitHub Actions check failures:
- Format code with cargo fmt (async_pattern_extraction.rs)
- Fix clippy::needless_range_loop warnings in edit_distance functions
- Fix clippy::unnecessary_cast in clustering.rs
- Fix clippy::useless_vec in decision_point.rs test
- Remove unused imports (Episode, PatternExtractionQueue)
- Prefix unused variable with underscore (_incomplete_id)

All clippy checks now pass with -D warnings

+++ memory-core/src/patterns/clustering.rs
@@ -230,7 +230,7 @@ impl PatternClusterer {
 
         let context_dist = 1.0 - self.context_distance(&episode.context, &centroid.context);
 
-        let steps_dist = (episode.steps.len() as f32 - centroid.avg_steps as f32).abs()
+        let steps_dist = (episode.steps.len() as f32 - centroid.avg_steps).abs()
             / centroid.avg_steps.max(1.0);
 
         let outcome_dist = if episode.outcome.is_some() == centroid.has_outcome {
